requirements.txt
numpy
jax
dm_haiku
optax
dm-reverb
rltools
